+++ bliss/flint/widgets/image_plot.py
@@ -16,6 +16,7 @@ import numpy
 
 from silx.gui import qt
 from silx.gui import icons
+from silx.gui import colors
 from silx.gui.plot.actions import histogram
 from silx.gui.plot.items.marker import Marker
 
@@ -140,6 +141,10 @@ class ImagePlotWidget(plot_helper.PlotWidget):
 
         self.__title = _Title(self.__plot)
 
+        self.__colormap = colors.Colormap("viridis")
+        """Each detector have a dedicated widget and a dedicated colormap"""
+        self.__colormapInitialized = False
+
         self.setFocusPolicy(qt.Qt.StrongFocus)
         self.__plot.installEventFilter(self)
         self.__plot.getWidgetHandle().installEventFilter(self)
@@ -154,7 +159,7 @@ class ImagePlotWidget(plot_helper.PlotWidget):
         toolBar = self.__createToolBar()
 
         # Try to improve the look and feel
-        # FIXME: THis should be done with stylesheet
+        # FIXME: This should be done with stylesheet
         line = qt.QFrame(self)
         line.setFrameShape(qt.QFrame.HLine)
         line.setFrameShadow(qt.QFrame.Sunken)
@@ -198,6 +203,31 @@ class ImagePlotWidget(plot_helper.PlotWidget):
         self.__plot.addItem(self.__tooltipManager.marker())
         self.__plot.addItem(self.__minMarker)
         self.__plot.addItem(self.__maxMarker)
+
+    def deviceName(self):
+        # FIXME: This have to be saved in the configuration
+        return self.windowTitle().split(" ")[0]
+
+    def configuration(self):
+        config = super(ImagePlotWidget, self).configuration()
+        try:
+            config.colormap = self.__colormap._toDict()
+        except Exception:
+            # As it relies on private API, make it safe
+            _logger.error("Impossible to save colormap preference", exc_info=True)
+        return config
+
+    def setConfiguration(self, config):
+        try:
+            self.__colormap._setFromDict(config.colormap)
+            self.__colormapInitialized = True
+        except Exception:
+            # As it relies on private API, make it safe
+            _logger.error("Impossible to restore colormap preference", exc_info=True)
+        super(ImagePlotWidget, self).setConfiguration(config)
+
+    def defaultColormap(self):
+        return self.__colormap
 
     def getRefreshManager(self) -> plot_helper.RefreshManager:
         return self.__refreshManager
@@ -298,6 +328,11 @@ class ImagePlotWidget(plot_helper.PlotWidget):
         self.__exportAction.setFlintModel(flintModel)
         self.__styleAction.setFlintModel(flintModel)
         self.__contrastAction.setFlintModel(flintModel)
+
+        if flintModel is not None:
+            if not self.__colormapInitialized:
+                style = flintModel.defaultImageStyle()
+                self.__colormap.setName(style.colormapLut)
 
     def setPlotModel(self, plotModel: plot_model.Plot):
         if self.__plotModel is not None:
@@ -589,7 +624,7 @@ class ImagePlotWidget(plot_helper.PlotWidget):
 
         legend = dataChannel.name()
         style = item.getStyle(self.__scan)
-        colormap = model_helper.getColormapFromItem(item, style)
+        colormap = model_helper.getColormapFromItem(item, style, self.__colormap)
 
         if style.symbolStyle is style_model.SymbolStyle.NO_SYMBOL:
             if image.ndim == 3: